+++ src/org-manager/containers/approve-organisation/approve-organisation.component.ts
@@ -5,42 +5,37 @@ import * as fromStore from '../../store/'
 import { Store, select } from '@ngrx/store';
 import {Observable, Subscription} from 'rxjs';
 import { OrganisationVM } from 'src/org-manager/models/organisation';
-import {take} from 'rxjs/operators';
+import {take, tap} from 'rxjs/operators';
 
 @Component({
     selector: 'app-org-pending-approve',
     templateUrl: './approve-organisation.component.html'
 })
-export class ApproveOrganisationComponent implements OnInit, OnDestroy {
+export class ApproveOrganisationComponent implements OnInit {
     orgForReview: OrganisationVM | null;
     serverResponseMessages$: Observable<any>;
+    disabled = true;
     constructor(
         public store: Store<fromOrganisationPendingStore.OrganisationRootState>
     ) { }
 
     ngOnInit() {
-        this.store.pipe(select(fromStore.getOrganisationForReview), take(1))
-            .subscribe((org: OrganisationVM) => {
-               if (!org) {
-                 this.store.dispatch(new fromRoot.Go({path: ['/pending-organisations']}));
-               }
-               this.orgForReview = org;
-            });
-
-        // Organisation error
-        // this.serverResponseMessages$ = this.store.pipe( select(fromStore.errorOrganisations));
-    }
-
-
-    onGoBack() {
-        this.store.dispatch(new fromRoot.Back());
+      this.store.pipe(select(fromStore.getOrganisationForReview), take(1)).subscribe((org: OrganisationVM) => {
+        if (!org) {
+          this.store.dispatch(new fromRoot.Go({path: ['/pending-organisations']}));
+        }
+        this.orgForReview = org;
+      });
+      this.serverResponseMessages$ = this.store.pipe(select(fromStore.getErrorMessage), tap(message => {
+        if (message) {
+          this.disabled = true;
+        }
+      }));
     }
 
     onApproveOrganisations() {
         this.store.dispatch(new fromOrganisationPendingStore.ApprovePendingOrganisations(this.orgForReview));
+        this.disabled = false;
     }
 
-    ngOnDestroy() {
-        // this.$reviewedOrganisationsSubscription.unsubscribe();
-    }
 }
