+++ src/org-manager/store/effects/index.ts
@@ -1,12 +1,9 @@
 import { OrganisationEffects } from './organisation.effects';
-import { PendingOrgEffects } from './organisation-pending.effects';
 
 export const effects: any[] = [
   OrganisationEffects,
-  PendingOrgEffects,
 ];
 
 export * from './organisation.effects';
-export * from './organisation-pending.effects';
